codeswitch/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-decouple>=3.8
Pillow>=10.0
gunicorn>=21.0
requests>=2.31
dj-database-url>=2.1
whitenoise>=6.6
django-csp>=3.8
django-axes>=6.4
